move the controller to root of controllers
